[#26] Use --no-pager when executing `git log`
By default git uses a "pager" to go through pages of data (usually
`less`).

In the case of using `git log | grep [...]` we were disabling the
paging behavior by "accident", not by design/intention.

The `--no-merges` does *not* change the output of `git log`, is the
lack of `| grep [...]`.

`git` has the option to exectute a command without paging, using
`--no-pager` flag.

https://git-scm.com/docs/git#Documentation/git.txt---no-pager