[#40] Show commits "chronologically" in Zulip messages
git log shows the "newest" commit at the top, and this is fine for a
familiar UI (with a pager), where the number of commits can be a lot.
When printing the commit messages in a chat message, it is easier to
understand the change with the commit messages ordered
"chronologically", rather than reverse "chronologically".

"newest" and "chronologically" are used loosely here -- they mean the
order in which they appear in the git branching tree, and don't really
have anything to do with the time when they were committed, etc.